# ApiGateway/app/openapi.yaml
openapi: 3.0.0
info:
  title: Chat Application API Gateway (SOA/CQRS)
  version: 1.0.0
  description: |
    API Gateway serving Chat Command, Chat Query, and Customer Management services.
    
    **NOTE:**
    1. Public Endpoints: `/auth/login`, `/auth/logout`, `/customer/create/account`, and `/customer/forgetPass/**`.
    2. Authenticated Endpoints: All other `/customer`, `/command`, and `/query` endpoints require the `jwt_token` cookie.
    3. **Internal Endpoints (Service-to-Service)**: Endpoints starting with `/internal/**` are for internal communication between services only and should not be called directly from client/frontend.
servers:
  - url: http://localhost:8080
    description: Nginx Gateway (Local Development)
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt_token
      description: JWT Token set after successful login.
  schemas:
    GenericResponse:
      type: object
      properties:
        success:
          type: boolean
          description: True if request is successful, False if there is a logical or business error.
        message:
          type: string
          description: Detailed notification about the result.
        data:
          type: object
          nullable: true
          description: Returned data (depends on the endpoint).
      required:
        - success
        - message
    
    # --- Auth & Customer Schemas ---
...
    OtpEmailRequest:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
      required:
        - email
    OtpEmailValidateRequest:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        otpCode:
          type: string
          example: '123456'
      required:
        - email
        - otpCode
    OtpGenerateResponseData:
      type: object
      properties:
        otpCode:
          type: string
          example: '123456'
        statusMessage:
          type: string
          example: New OTP generated successfully.
    SendMailRequest:
      type: object
      properties:
        to:
          type: string
          example: recipient@example.com
        subject:
          type: string
          example: Mã OTP của bạn
        body:
          type: string
          example: Mã OTP là 123456. Mã này có hiệu lực trong 5 phút.
      required:
        - to
        - subject
        - body

    # --- Chat Command Schemas ---
...
paths:
...
  /internal/generate/email:
    post:
      tags:
        - OTP Service (Internal Use Only)
      summary: Generate OTP (Internal Use Only)
      operationId: generateOtpByEmailInternal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpEmailRequest'
      responses:
        '200':
          description: OTP generated successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OtpGenerateResponseData'
        '400':
          description: Invalid request.
        '500':
          description: Internal server error.
  /internal/resend/email:
    post:
      tags:
        - OTP Service (Internal Use Only)
      summary: Resend OTP (Internal Use Only)
      operationId: resendOtpByEmailInternal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpEmailRequest'
      responses:
        '200':
          description: OTP resent successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OtpGenerateResponseData'
        '400':
          description: Invalid request.
        '500':
          description: Internal server error.
  /internal/validate/email:
    post:
      tags:
        - OTP Service (Internal Use Only)
      summary: Validate OTP (Internal Use Only)
      operationId: validateOtpByEmailInternal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpEmailValidateRequest'
      responses:
        '200':
          description: OTP validated successfully or not.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                success: true
                message: OTP validated successfully.
                data: null
        '400':
          description: Invalid request.
        '500':
          description: Internal server error.
  # --- Mail Service Internal Endpoint ---
  /internal/send:
    post:
      tags:
        - Mail Service (Internal Use Only)
      summary: Send a simple email (Internal Use Only)
      operationId: sendMailInternal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMailRequest'
      responses:
        '200':
          description: Email sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                success: true
                message: Email sent successfully.
                data: null
        '400':
          description: Invalid request.
        '500':
          description: Failed to send email.